Refactor, renaming, moved setting number of evaluations to Hooke class
Number of evaluations which is Hooke class variable was unnecesarily changed in Rosenbrock and Woods classes. Which meant anytime somebody wanted to find minimum of new function and have information about function evaluations had to take care of it inside method calculating values of meathematical objective function now this part is moved to Hooke class which is class of Search Method.